Schedule INKContInternal destruction on local queue
This event is being scheduled on the current `EThread`, so the
signalling schedule call is unnecessary. Use the non-signalling `_local`
function instead.

This has an additional benefit: This allows `Dedicated` `EThreads` to
destroy continuations. `schedule_imm` asserts if called from a
`Dedicated` `EThread`. `schedule_imm_local` checks for this condition
and schedules the event on `ET_CALL` if necessary.